Replace maplit with ahash for improved performance
The existing maplit library for hashmap and hashset initialization was replaced with ahash, which is typically faster and more memory efficient. Additionally, some also transitioned from the the now deprecated lazy_static! to just lazy_static. The implementation of various caching functions was updated to use SizedCache instead of LruCache to improve speed.