avoid additional payload copy per matching charset
Just make a single copy of the payload per guess, rather than 1 per
candidate match.

Ideally we'd use a lifetime to reference the input payload,
but the submatch stuff makes that difficult to wrangle,
so we satisfy ourselves with a Cow.

Since this commit changes the published API, bump the version
number in the manifest.