from_bytes is now fallible
My immediate goal is to avoid the unwrap() panic when an invalid
encoding name is passed to the library, as this is highly undesirable in
a highly stateful service application.

This commit replaces some bare `.unwrap()` calls with either error
message propagation or `.expect(REASON)` to indicate why the panic
can never happen at runtime.