impl Ord for CharsetMatch
Define a total order via OrderedFloat; this allows the replacement
of a somewhat sketchy `partial_cmp().unwrap()` with `cmp()`, reducing
the number of places in this crate that might panic.

Part of this commit makes CoherenceMatches have pub(crate) visibility;
it appears not to part of any public API, and this commit changes
the type of one its fields, so this commit is an API change anyway.