introduce own Encoding type
This was intended to be essentially a refactory commit that moves
away from operating on encoding names and free functions that
call into the unmaintained encoding crate, instead towards a
type that can then introduce calls to the successor encoding_rs
crate.

I'd intended that no functional changes occur as part of this
change, but some of the largesets tests started to fail: on closer
inspection, there were a number of utf-8 files that were placed
in the ascii directory: they were misclassified and the previous
logic considered them to be validly ascii when they were actually
utf-8.

I don't see precisely where this change in behavior originates from, but
I suspect that it is to do with some name canonicalization.  My
resolution here is to move those test files to the utf-8 directory as
they are now correctly classified.

The material changes in this commit are the definitions of the various
Encoding instances and the corresponding alias maps.

The rest of the changes are largely mechanical consequence of moving
from strings to &Encoding.

refs: https://github.com/nickspring/charset-normalizer-rs/issues/38