switch to encoding_rs as the backend implementation
Adds encoding_rs, but does not remove encoding.  Both implementations
exist in the code and the results of both can be compared by changing an
`if false` to `if true`.

When I do that, a single windows-1255 test case has a discrepancy,
which is due to encoding_rs's implementation accepting the input
file.

I think that this is acceptable and reasonable.

I'll remove encoding in a follow up commit.

refs: https://github.com/nickspring/charset-normalizer-rs/issues/38