mess detector: switch to fixed size cache
Previously, the mess detector used an unbounded cache.

Unbounded cache size is undesirable as it presents as a potential attack
surface if the detector is operating on data coming in from the network:
an attacker could generate poison content consisting of all possible
unicode codepoints across a series of otherwise innocuous messages to
gradually consume the memory on the system.

This commit switches it to be a fixed size cache, equal to the initial
reserved capacity that was in use prior to this change.