limit memory usage when WantDecode::No is in use
Previously, we'd always materialize the decoded string, even if we knew
that we didn't need the result.

This commit adjusts to use the more complex streaming decode api from
encoding_rs so that we can set an upper bound on memory usage when
operating in validation only mode.

FWIW, this change has no measurable impact on the results of
running the performance tests.